Move PreAnalyzer from SQL to QL
