+++ x-pack/plugin/sql/src/main/java/org/elasticsearch/xpack/sql/session/SqlSession.java
@@ -11,6 +11,9 @@ import org.elasticsearch.action.ActionListener;
 import org.elasticsearch.client.internal.Client;
 import org.elasticsearch.client.internal.ParentTaskAssigningClient;
 import org.elasticsearch.tasks.TaskCancelledException;
+import org.elasticsearch.xpack.ql.analyzer.PreAnalyzer;
+import org.elasticsearch.xpack.ql.analyzer.PreAnalyzer.PreAnalysis;
+import org.elasticsearch.xpack.ql.analyzer.TableInfo;
 import org.elasticsearch.xpack.ql.expression.function.FunctionRegistry;
 import org.elasticsearch.xpack.ql.index.IndexCompatibility;
 import org.elasticsearch.xpack.ql.index.IndexResolution;
@@ -21,9 +24,6 @@ import org.elasticsearch.xpack.ql.plan.logical.LogicalPlan;
 import org.elasticsearch.xpack.ql.rule.RuleExecutor;
 import org.elasticsearch.xpack.sql.analysis.analyzer.Analyzer;
 import org.elasticsearch.xpack.sql.analysis.analyzer.AnalyzerContext;
-import org.elasticsearch.xpack.sql.analysis.analyzer.PreAnalyzer;
-import org.elasticsearch.xpack.sql.analysis.analyzer.PreAnalyzer.PreAnalysis;
-import org.elasticsearch.xpack.sql.analysis.analyzer.TableInfo;
 import org.elasticsearch.xpack.sql.analysis.analyzer.Verifier;
 import org.elasticsearch.xpack.sql.execution.PlanExecutor;
 import org.elasticsearch.xpack.sql.optimizer.Optimizer;
